requirements.txt
openai>=1.0.0
cryptography>=41.0.0
pyasn>=1.6.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
